Add a Thrift parser in wasm plugin
In our theoretical example, we have the information we need from a
static source - the Thrift file.

If the Thrift file itself doesn't hit contain the requested scopes,
then we don't even need to hit the auth service.

From a reference point-of-view, both are valuable implementations to
look at. Wanted to add this one.

At this very moment, the implementation still calls out to the auth
service.